+++ app/components/filesUpload/FilesUpload.tsx
@@ -3,7 +3,7 @@ import classNames from "classnames";
 import { useActionData } from "react-router";
 import { type ErrorMessageProps } from "~/components/inputs";
 import InputError from "~/components/inputs/InputError";
-import { type Context } from "~/domains/contexts";
+import { type Context } from "~/domains/userData";
 import { useJsAvailable } from "~/services/useJsAvailable";
 import {
   errorStyling,
