+++ app/components/form/ValidatedFlowForm.tsx
@@ -2,7 +2,7 @@ import { ValidatedForm } from "@rvf/react-router";
 import { useLocation } from "react-router";
 import type { ButtonNavigationProps } from "~/components/form/ButtonNavigation";
 import { ButtonNavigation } from "~/components/form/ButtonNavigation";
-import type { Context } from "~/domains/contexts";
+import type { Context } from "~/domains/userData";
 import { StrapiFormComponents } from "~/services/cms/components/StrapiFormComponents";
 import type { StrapiFormComponent } from "~/services/cms/models/StrapiFormComponent";
 import { CSRFKey } from "~/services/security/csrf/csrfKey";
