+++ app/domains/__test__/TestCases.ts
@@ -1,5 +1,5 @@
 import type { BeratungshilfeFormularUserData } from "~/domains/beratungshilfe/formular";
-import type { BeratungshilfeVorabcheckContext } from "~/domains/beratungshilfe/vorabcheck/context";
+import type { BeratungshilfeVorabcheckUserData } from "~/domains/beratungshilfe/vorabcheck/userData";
 import type { FluggastrechteUserData } from "~/domains/fluggastrechte/formular/userData";
 import type { FluggastrechtVorabcheckContext } from "~/domains/fluggastrechte/vorabcheck/context";
 import type { GeldEinklagenVorabcheckContext } from "~/domains/geldEinklagen/vorabcheck/context";
@@ -8,7 +8,7 @@ import type { ProzesskostenhilfeFormularContext } from "../prozesskostenhilfe/fo
 
 export type TestCases<
   T extends
-    | BeratungshilfeVorabcheckContext
+    | BeratungshilfeVorabcheckUserData
     | BeratungshilfeFormularUserData
     | GeldEinklagenVorabcheckContext
     | FluggastrechtVorabcheckContext
