+++ app/domains/flows.server.ts
@@ -8,10 +8,10 @@ import { geldEinklagenVorabcheck } from "~/domains/geldEinklagen/vorabcheck";
 import type { Config } from "~/services/flow/server/buildFlowController";
 import type { FlowTransitionConfig } from "~/services/flow/server/flowTransitionValidation";
 import type { Replacements } from "~/util/applyStringReplacement";
-import type { Context } from "./contexts";
 import type { FlowId } from "./flowIds";
 import type { Guards } from "./guards.server";
 import { prozesskostenhilfeFormular } from "./prozesskostenhilfe/formular";
+import type { Context } from "./userData";
 
 type FlowMigration = {
   source: FlowId;
