+++ app/services/array/index.ts
@@ -1,5 +1,5 @@
 import type { AllContextKeys } from "~/domains/common";
-import type { Context } from "~/domains/contexts";
+import type { Context } from "~/domains/userData";
 
 export const arrayChar = "#";
 export const splitArrayName = (key: string) => key.split(arrayChar);
