+++ app/services/flow/formular/buildFormularServerTranslations.ts
@@ -1,5 +1,5 @@
-import type { Context } from "~/domains/contexts";
 import type { Flow } from "~/domains/flows.server";
+import type { Context } from "~/domains/userData";
 import { getArraySummaryPageTranslations } from "~/services/array/getArraySummaryPageTranslations";
 import type { StrapiFormFlowPage } from "~/services/cms/models/StrapiFormFlowPage";
 import type { Translations } from "~/services/translations/getTranslationByKey";
