+++ app/services/flow/pageData.ts
@@ -1,5 +1,5 @@
 import merge from "lodash/merge";
-import type { Context } from "~/domains/contexts";
+import type { Context } from "~/domains/userData";
 import { type PageData } from "./pageDataSchema";
 
 export function addPageDataToUserData(userData: Context, pageData: PageData) {
