+++ app/services/session.server/arrayDeletion.ts
@@ -1,6 +1,6 @@
 import type { Session } from "react-router";
-import type { Context, ObjectType } from "~/domains/contexts";
 import { type FlowId } from "~/domains/flowIds";
+import type { Context, ObjectType } from "~/domains/userData";
 import { getSessionManager, updateSession } from "~/services/session.server";
 import { filterFormData } from "~/util/filterFormData";
