+++ app/services/session.server/index.ts
@@ -3,8 +3,8 @@ import { type MergeWithCustomizer } from "lodash";
 import mergeWith from "lodash/mergeWith";
 import type { Cookie, Session } from "react-router";
 import { createSessionStorage, createCookie } from "react-router";
-import { type Context } from "~/domains/contexts";
 import { flowIds, type FlowId } from "~/domains/flowIds";
+import { type Context } from "~/domains/userData";
 import { config } from "~/services/env/env.server";
 import { useSecureCookie } from "~/util/useSecureCookie";
 import {
