+++ app/services/upload/buildFileUploadError.ts
@@ -1,5 +1,5 @@
 import { validationError, type ValidationResult } from "@rvf/react-router";
-import { type Context } from "~/domains/contexts";
+import { type Context } from "~/domains/userData";
 
 /**
  * Need to remove the validation error's object notation to conform to what the frontend expects
