+++ app/services/upload/validateUploadedFile.ts
@@ -2,7 +2,7 @@
 import { type ValidationResult } from "@rvf/react-router";
 import { withZod } from "@rvf/zod";
 import { z, type ZodTypeAny } from "zod";
-import { type Context } from "~/domains/contexts";
+import { type Context } from "~/domains/userData";
 import { type PDFFileMetadata } from "~/util/file/pdfFileSchema";
 
 export async function validateUploadedFile(
