+++ app/services/validation/validateFormData.server.ts
@@ -1,5 +1,5 @@
 import { parseFormData } from "@rvf/react-router";
-import { type Context } from "~/domains/contexts";
+import { type Context } from "~/domains/userData";
 import { schemaForFieldNames } from "./stepValidator/schemaForFieldNames";
 
 export async function validateFormData(pathname: string, formData: FormData) {
